+++ common/persistence/sql/workflowStateMaps.go
@@ -55,7 +55,7 @@ func updateActivityInfos(
 			}
 
 			rows[i] = sqlplugin.ActivityInfoMapsRow{
-				ShardID:      int64(shardID),
+				ShardID:      int32(shardID),
 				NamespaceID:  namespaceID,
 				WorkflowID:   workflowID,
 				RunID:        runID,
@@ -73,7 +73,7 @@ func updateActivityInfos(
 	if len(deleteInfos) > 0 {
 		for _, v := range deleteInfos {
 			result, err := tx.DeleteFromActivityInfoMaps(&sqlplugin.ActivityInfoMapsFilter{
-				ShardID:     int64(shardID),
+				ShardID:     int32(shardID),
 				NamespaceID: namespaceID,
 				WorkflowID:  workflowID,
 				RunID:       runID,
@@ -104,7 +104,7 @@ func getActivityInfoMap(
 ) (map[int64]*persistenceblobs.ActivityInfo, error) {
 
 	rows, err := db.SelectFromActivityInfoMaps(&sqlplugin.ActivityInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
@@ -134,7 +134,7 @@ func deleteActivityInfoMap(
 ) error {
 
 	if _, err := tx.DeleteFromActivityInfoMaps(&sqlplugin.ActivityInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
@@ -162,7 +162,7 @@ func updateTimerInfos(
 				return err
 			}
 			rows[i] = sqlplugin.TimerInfoMapsRow{
-				ShardID:      int64(shardID),
+				ShardID:      int32(shardID),
 				NamespaceID:  namespaceID,
 				WorkflowID:   workflowID,
 				RunID:        runID,
@@ -178,7 +178,7 @@ func updateTimerInfos(
 	if len(deleteInfos) > 0 {
 		for _, v := range deleteInfos {
 			result, err := tx.DeleteFromTimerInfoMaps(&sqlplugin.TimerInfoMapsFilter{
-				ShardID:     int64(shardID),
+				ShardID:     int32(shardID),
 				NamespaceID: namespaceID,
 				WorkflowID:  workflowID,
 				RunID:       runID,
@@ -208,7 +208,7 @@ func getTimerInfoMap(
 ) (map[string]*persistenceblobs.TimerInfo, error) {
 
 	rows, err := db.SelectFromTimerInfoMaps(&sqlplugin.TimerInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
@@ -237,7 +237,7 @@ func deleteTimerInfoMap(
 ) error {
 
 	if _, err := tx.DeleteFromTimerInfoMaps(&sqlplugin.TimerInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
@@ -265,7 +265,7 @@ func updateChildExecutionInfos(
 				return err
 			}
 			rows[i] = sqlplugin.ChildExecutionInfoMapsRow{
-				ShardID:      int64(shardID),
+				ShardID:      int32(shardID),
 				NamespaceID:  namespaceID,
 				WorkflowID:   workflowID,
 				RunID:        runID,
@@ -280,7 +280,7 @@ func updateChildExecutionInfos(
 	}
 	if deleteInfos != nil {
 		if _, err := tx.DeleteFromChildExecutionInfoMaps(&sqlplugin.ChildExecutionInfoMapsFilter{
-			ShardID:     int64(shardID),
+			ShardID:     int32(shardID),
 			NamespaceID: namespaceID,
 			WorkflowID:  workflowID,
 			RunID:       runID,
@@ -302,7 +302,7 @@ func getChildExecutionInfoMap(
 ) (map[int64]*persistenceblobs.ChildExecutionInfo, error) {
 
 	rows, err := db.SelectFromChildExecutionInfoMaps(&sqlplugin.ChildExecutionInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
@@ -332,7 +332,7 @@ func deleteChildExecutionInfoMap(
 ) error {
 
 	if _, err := tx.DeleteFromChildExecutionInfoMaps(&sqlplugin.ChildExecutionInfoMapsFilter{
-		ShardID:     int64(shardID),
+		ShardID:     int32(shardID),
 		NamespaceID: namespaceID,
 		WorkflowID:  workflowID,
 		RunID:       runID,
